+++ tests/testthat/setup.R
@@ -6,16 +6,27 @@ withr::local_options(
   .local_envir = teardown_env()
 )
 
+testing_env_variables <- c(
+  "SHAREPOINT_SITE_URL"
+)
+
+# Utility function for skipping tests if ENV variables are not set
+skip_offline_test <- function() {
+  skip_if(!all(testing_env_variables %in% names(Sys.getenv())), message = NULL)
+}
+
 # just to have clean messages in tests
 quiet_connect <- function(...) {
   suppressMessages(connector_sharepoint(...))
 }
 
-my_site <- Sys.getenv("SHAREPOINT_SITE_URL")
-
-if (my_site == "") {
-  cli::cli_alert_info("No site found, skipping tests.")
+if (!all(testing_env_variables %in% names(Sys.getenv()))) {
+  cli::cli_alert_warning(
+    "Not all testing parameters are set. Please set environment variable
+      SHAREPOINT_SITE_URL in order to be able to test the whole package."
+  )
 } else {
-  cli::cli_alert_info("Using {my_site} for tests")
-  setup_connector <- quiet_connect(site_url = my_site)
+  setup_site_url <- Sys.getenv("SHAREPOINT_SITE_URL")
+  cli::cli_alert_info("Using {setup_site_url} for tests")
+  setup_connector <- quiet_connect(site_url = setup_site_url)
 }